requirements.txt
pymupdf==1.23.22
pytesseract==0.3.10 # requires Tesseract OCR to also be installed
pillow==10.2.0
pandas==2.2.2
